spliting components to seperate file from header
