+++ src/view/header/Header.tsx
@@ -1,83 +1,11 @@
 import React, { useState } from "react";
 import { Direction } from "../../logic/Dice/dice";
 import { MachineContext } from "../../logic/Game/stateMachine";
+import { DirectionContainer, StepContainer } from "./Containers";
+import { RollDiceButton } from "./Containers";
 interface HeaderProps {
   rollDice: () => MachineContext;
 }
-
-interface RollDiceButtonProps {
-  updateState: () => void;
-}
-
-interface DirectionBoxProps {
-  number: number;
-  direction: Direction;
-}
-
-const BOX_MARGIN = "10px";
-
-const containerStyle = {
-  backgroundColor: "white",
-  display: "flex",
-  border: "2px rgb(240,240,240) solid",
-  margin: BOX_MARGIN,
-  borderRadius: "5%",
-  gap: "10px",
-  padding: "10px",
-  alignItems: "center",
-  height: "69px",
-};
-
-const boxStyle = {
-  width: "30px",
-  height: "30px",
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-  border: "2px solid",
-};
-const RollDiceButton: React.FC<RollDiceButtonProps> = ({ updateState }) => {
-  return (
-    <button
-      style={{
-        backgroundColor: "black",
-        cursor: "pointer",
-        width: "182px",
-        height: "48px",
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-      }}
-      onClick={() => updateState()}
-    >
-      <h4 style={{ color: "white" }}>ROLL DICE</h4>
-    </button>
-  );
-};
-const DirectionContainer: React.FC<DirectionBoxProps> = ({
-  number,
-  direction,
-}) => {
-  return (
-    <div
-      style={{
-        ...containerStyle,
-      }}
-    >
-      <p>DIRECTION</p>
-      <div style={{ ...boxStyle }}>{number}</div>
-      <h4>{direction}</h4>
-    </div>
-  );
-};
-
-const StepContainer: React.FC<{ number: number }> = ({ number }) => {
-  return (
-    <div style={{ ...containerStyle }}>
-      <p>STEPS:</p> <div style={{ ...boxStyle }}>{number}</div>
-    </div>
-  );
-};
 
 export const Header: React.FC<HeaderProps> = ({ rollDice }) => {
   const [direction, setDirection] = useState<Direction>(Direction.North);
@@ -86,9 +14,12 @@ export const Header: React.FC<HeaderProps> = ({ rollDice }) => {
 
   const updateState = () => {
     const context = rollDice();
-    setDirection(context.userLastTurn);
-    setDirectionNumber(context.directionNumber);
-    setStep(context.step);
+    console.log(context, "update state");
+    if (context.userLastTurn) {
+      setDirection(context.userLastTurn);
+      setDirectionNumber(context.directionNumber);
+      setStep(context.step);
+    }
   };
   return (
     <div
